Script_for_serverless/requirements.txt
fastapi
uvicorn
tflite-runtime
pillow
numpy
python-multipart
